Refactor breadcrumbs and show patch title on task page
