+++ src/hooks/index.ts
@@ -22,4 +22,3 @@ export { useSpruceConfig } from "./useSpruceConfig";
 export { useUserSettings } from "./useUserSettings";
 export { useDateFormat } from "./useDateFormat";
 export { useBreadcrumbRoot } from "./useBreadcrumbRoot";
-export { useBreadcrumbVersion } from "./useBreadcrumbVersion";
